Feat: Relation C,D API Error 처리
- 공통으로 사용할 Error Handler 작성
- 각 domain별 NotFoundException 작성
- Relation Create시에 해당 Node가 없을 때 Error 처리
- Relation Delete시에 해당 Node가 없거나 해당 Relation이 없을 때 Error 처리

+++ src/domain/relation/relation.service.ts
@@ -1,23 +1,70 @@
 import { Injectable } from "@nestjs/common";
+import { AlbumRepository } from "../album/album.repository";
+import { NotFoundAlbumException } from "../album/exception/NotFoundAlbumException";
+import { NotFoundMusicianException } from "../musician/exception/NotFoundMusicianException";
+import { MusicianRepository } from "../musician/musician.repository";
+import { NotFoundSongException } from "../song/exception/NotFoundSongException";
+import { SongRepository } from "../song/song.repository";
+import { NotFoundRelationException } from "./exception/NotFoundRelationException";
 import { RelationRepository } from "./relation.repository";
 
 @Injectable()
 export class RelationService {
-	constructor(private relationRepository: RelationRepository) {}
+	constructor(
+		private relationRepository: RelationRepository,
+		private musicianRepository: MusicianRepository,
+		private songRepository: SongRepository,
+		private albumRepository: AlbumRepository
+	) {}
 
-	createMSong(body) {
+	async createMSong(body) {
+		// find musician
+		const musician = await this.musicianRepository.findOne(body.musicianId);
+		if (musician === undefined) throw new NotFoundMusicianException();
+		// find song
+		const song = await this.songRepository.findOne(body.songId);
+		if (song === undefined) throw new NotFoundSongException();
+		// create relation
 		return this.relationRepository.createMSong(body);
 	}
 
-	createASong(body) {
+	async createASong(body) {
+		// find album
+		const album = await this.albumRepository.findOne(body.albumId);
+		if (album === undefined) throw new NotFoundAlbumException();
+		// find song
+		const song = await this.songRepository.findOne(body.songId);
+		if (song === undefined) throw new NotFoundSongException();
+		// create relation
 		return this.relationRepository.createASong(body);
 	}
 
-	deleteMSong(musicianId, songId) {
-		return this.relationRepository.deleteMSong(musicianId, songId);
+	async deleteMSong(musicianId, songId) {
+		// find musician
+		const musician = await this.musicianRepository.findOne(musicianId);
+		if (musician === undefined) throw new NotFoundMusicianException();
+		// find song
+		const song = await this.songRepository.findOne(songId);
+		if (song === undefined) throw new NotFoundSongException();
+		// delete relation
+		const result = await this.relationRepository.deleteMSong(
+			musicianId,
+			songId
+		);
+		if (result === undefined) throw new NotFoundRelationException();
+		return result;
 	}
 
-	deleteASong(albumId, songId) {
-		return this.relationRepository.deleteASong(albumId, songId);
+	async deleteASong(albumId, songId) {
+		// find album
+		const album = await this.albumRepository.findOne(albumId);
+		if (album === undefined) throw new NotFoundAlbumException();
+		// find song
+		const song = await this.songRepository.findOne(songId);
+		if (song === undefined) throw new NotFoundSongException();
+		// delete relation
+		const result = this.relationRepository.deleteASong(albumId, songId);
+		if (result === undefined) throw new NotFoundRelationException();
+		return result;
 	}
 }
